refactor: typescript migration, components simplifying
