+++ components/Layout/ServiceStatus.tsx
@@ -2,27 +2,28 @@ import { useDispatch } from 'react-redux';
 
 import { ServiceStatusInfo } from '@autonolas/frontend-library';
 
-import { getLatestMintedNft, getLeaderboardList } from 'common-util/api';
+import { getLeaderboardList } from 'common-util/api';
 import { useHealthCheckup } from 'common-util/hooks/useHealthCheckup';
-import { useHelpers } from 'common-util/hooks/useHelpers';
-import { setLeaderboard, setNftDetails } from 'store/setup';
+import { setLeaderboard } from 'store/setup';
 
 const MINUTE = 60 * 1000;
 
-const ServiceStatus = () => {
+export const ServiceStatus = () => {
   const dispatch = useDispatch();
-  const { account } = useHelpers();
+  // const { account } = useHelpers();
 
   const pollingCallback = async () => {
     // fetch leaderboard list
     const list = await getLeaderboardList();
     dispatch(setLeaderboard(list));
 
+    // NOTE: leaderboard is now the default page thus no need to poll minted nft
+    //
     // update badge if the user is logged-in
-    if (account) {
-      const { details, tokenId } = await getLatestMintedNft(account);
-      dispatch(setNftDetails({ tokenId, ...(details || {}) }));
-    }
+    // if (account) {
+    //   const { details, tokenId } = await getLatestMintedNft(account);
+    //   dispatch(setNftDetails({ tokenId, ...(details || {}) }));
+    // }
   };
 
   const [isHealthy] = useHealthCheckup(
@@ -33,5 +34,3 @@ const ServiceStatus = () => {
 
   return <ServiceStatusInfo isHealthy={isHealthy} appType="iekit" />;
 };
-
-export default ServiceStatus;
