--- components/Leaderboard/MintNft/Education/data.json
@@ -1,20 +1,0 @@
-[
-  {
-    "id": "9b939392-4b49-49de-87de-16d0ed479597",
-    "slug": "leaderboard",
-    "text": "Climb the leaderboard by completing actions that contribute to Autonolas’ success.",
-    "docsLink": "#leaderboard"
-  },
-  {
-    "id": "cb482852-26a4-4e52-bb9d-2c1b05cb5833",
-    "slug": "actions",
-    "text": "Complete actions to earn points, climb the leaderboard and upgrade your badge.",
-    "docsLink": "#actions"
-  },
-  {
-    "id": "78bcf070-d8ad-4a11-8d0d-bb746525b075",
-    "slug": "badge",
-    "text": "Show off your leaderboard rank and promote Autonolas with a badge that evolves as you contribute.",
-    "docsLink": "#badge"
-  }
-]
