+++ components/Leaderboard/styles.tsx
