--- pages/leaderboard.jsx
@@ -1,3 +1,0 @@
-import Leaderboard from 'components/Leaderboard';
-
-export default Leaderboard;
