+++ pages/leaderboard.tsx
@@ -1,0 +1,3 @@
+import { Leaderboard } from 'components/Leaderboard';
+
+export default Leaderboard;
