+++ pages/staking.tsx
