--- util/education.js
@@ -1,6 +1,0 @@
-import data from 'common-util/Education/data.json';
-
-export const getEducationItemByComponent = (component) => {
-  const items = data.filter((item) => component === item.component);
-  return items[0];
-};
